+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+<<<<<<< HEAD
+- New component nc-button #17
 - New component nc-slideshow #6
 - New component nc-modal #4
 - New component nc-list #11